==> analysis-tools/06_car_traffic_comp.R <==
## Car traffic

# Reading input car traffic data

amCarObservations <- read_sf("./data/observationsJoined/carAmObsJoined2Network.sqlite") %>% 
  dplyr::select(link_id, starts_with("x")) 
pmCarObservations <- read_sf("./data/observationsJoined/carPmObsJoined2Network.sqlite") %>% 
  dplyr::select(link_id, starts_with("x"))

if (all(colnames(amCarObservations)==colnames(pmCarObservations))) {
  if (nrow(filter(amCarObservations,link_id%in%pmCarObservations$link_id))==0) {
    carObservations <- rbind(st_drop_geometry(amCarObservations), 
                             st_drop_geometry(pmCarObservations))
  }else print("There is a duplicate link_id in AM and PM data!")
}else print("Colnames in AM and PM car observations does not match!")

carSimulation <- read_sf("./data/simOutputJoined/networkLinksWithHourlyVol.sqlite", 
                         layer="car")

# Restructuring observation and simulation car data}

carObservationsLong <- carObservations %>% 
  pivot_longer(cols = starts_with("x"), names_to="hour", values_to="obs_volume") %>% 
  mutate(hour=as.integer(str_remove(hour,"x"))) #%>% 
# mutate(Source="Observation")

carSimulationFiltered <- carSimulation %>% 
  st_drop_geometry() %>% 
  dplyr::select(link_id=id, starts_with("X")) %>% 
  filter(link_id %in% carObservations$link_id)

carSimulationLong <- carSimulationFiltered %>% 
  pivot_longer(cols = starts_with("X"), names_to="hour", values_to="sim_volume") %>% 
  mutate(hour=as.double(str_remove(hour,"X"))) %>% 
  mutate(sim_volume=sim_volume*10)  %>% # Because I used 10% sample 
  mutate(hour=case_when(hour==24 ~ 0,
                        hour==25 ~ 1,
                        hour==26 ~ 2,
                        hour==27 ~ 3,
                        hour==28 ~ 4,
                        hour==29 ~ 5,
                        hour==30 ~ 6,
                        TRUE ~ hour)) %>% 
  group_by(link_id, hour) %>% 
  summarise(sim_volume=sum(sim_volume,na.rm = T))
# mutate(Source="Simulation")

carData <- carObservationsLong %>% 
  left_join(carSimulationLong, by = c("link_id","hour"))

### Plotting the volumes

carSimulationFiltered <- carSimulation %>% 
  filter(total_vol>0) %>% 
  mutate(volumeAdjusted=total_vol*10) %>% 
  dplyr::select(volumeAdjusted)

carBreakPoints <- pretty(carSimulationFiltered$volumeAdjusted, n = 6)

ggplot(carSimulationFiltered) +
  annotation_map_tile(type="osmgrayscale",zoom=9, alpha=0.6) +
  geom_sf(aes(size=volumeAdjusted, colour=volumeAdjusted, alpha=volumeAdjusted)) +
  # geom_sf(aes( size=n, colour=n ))  +
  scale_color_viridis(option="magma", trans="sqrt", breaks=carBreakPoints, name="Daily volume" ) +
  scale_alpha_continuous(name="Daily volume", trans="sqrt", range=c(0.05, .9), breaks=carBreakPoints) +
  scale_size_continuous(name="Daily volume", range=c(0.01, 2), breaks=carBreakPoints) +
  theme_void() + 
  guides( colour = guide_legend()) +
  theme(
    legend.position = c(0.85, 0.8),
    text = element_text(color = "#22211d"),
    plot.background = element_rect(fill = "#f5f5f2", color = NA), 
    panel.background = element_rect(fill = "#f5f5f2", color = NA), 
    legend.background = element_rect(fill = "#f5f5f2", color = NA)
  )

ggsave("carDailyTraffic.png",
       width = 7, height = 7)

carData %>% 
  # filter(link_id==11169) %>% 
  ggplot(aes(x=hour)) +
  geom_line(aes(y=obs_volume, color="observation")) + 
  geom_line(aes(y=sim_volume, color="simulation")) +
  facet_wrap("link_id")

ggsave(paste0(outputDir,"car.png"), width= 50, height = 50, units="cm")

# All car on the road traffice

carSimHourlyAll <- carSimulationLong %>% 
  filter(!is.na(sim_volume)) %>% 
  filter(hour<24) %>% 
  group_by(hour) %>% 
  summarise(volume=sum(sim_volume)) %>%
  mutate(volume_ratio=100*volume/sum(volume)) %>% 
  mutate(source="Simulation")

carObsHourlyAll <- carObservationsLong %>% 
  group_by(hour) %>% 
  summarise(volume=sum(obs_volume)) %>% 
  mutate(volume_ratio=100*volume/sum(volume)) %>% 
  mutate(source="Observation")

rbind(carObsHourlyAll,
      carSimHourlyAll) %>% 
  ggplot(aes(x=hour,y=volume_ratio, color=source)) +
  geom_col(aes(fill=source),position = "identity" , alpha=0.2) +
  geom_line() + 
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(color = "#22211d"),
    plot.background = element_rect(fill = "#f5f5f2", color = NA), 
    panel.background = element_rect(fill = "#f5f5f2", color = NA), 
    legend.background = element_rect(fill = "#f5f5f2", color = NA)
  ) +
  xlab("Hour of the day") +
  ylab("Traffic volume (%)")

ggsave("carTrafficOveral.png",
       width= 5, height = 3)

# Calculating GEH Statistic for car traffic

carDataWithGEH <- carData %>% 
  mutate(gehStat= sqrt((2*(sim_volume-obs_volume)^2)/(sim_volume+obs_volume)))

st_write(carDataWithGEH, paste0(outputDir,"calibrationData.sqlite"), 
         layer="car", delete_layer=T)

# Plotting the GEH for car traffic

carDataWithGEH %>% 
  ggplot(aes(x=hour)) +
  geom_line(aes(y=gehStat)) + 
  facet_wrap("link_id")

ggsave(paste0(outputDir,"carGeh.png"), width= 50, height = 50, units="cm")

carDataWithGEH %>% 
  drop_na(gehStat) %>% 
  group_by(link_id) %>% 
  summarise(avgGEH=mean(gehStat))

### Using precentage error

carDataWithGEH <- carData %>% 
  mutate(gehStat= sqrt((2*(sim_volume-obs_volume)^2)/(sim_volume+obs_volume))) %>% 
  drop_na(gehStat)

carDataWithGEH %>% 
  ggplot()+
  geom_histogram(aes(x=gehStat))

carDataWithApe <- carData %>% 
  mutate(ape= abs(obs_volume-sim_volume)/obs_volume )

carDataWithApe %>% 
  ggplot(aes(x=hour)) +
  geom_line(aes(y=ape)) + 
  facet_wrap("link_id")
ggsave(paste0(outputDir,"car_ape.png"), width= 50, height = 50, units="cm")

### Comparing hourly shares

carDataPct.h <- carData %>% 
  mutate(obs_volume=replace_na(obs_volume,0),
         sim_volume=replace_na(sim_volume,0)) %>% 
  group_by(link_id) %>% 
  mutate(obs.pct.h=obs_volume/sum(obs_volume, na.rm = T),
         sim.pct.h=sim_volume/sum(sim_volume, na.rm = T)) 

carDataPct.h %>% 
  ggplot(aes(x=hour))+
  geom_line(aes(y=obs.pct.h, color="observation")) + 
  geom_line(aes(y=sim.pct.h, color="simulation")) +
  facet_wrap("link_id")

ggsave(paste0(outputDir,"carpct.png"), width= 50, height = 50, units="cm")

# Hourly aggregated WAPE

error_wape <- carDataPct.h %>% 
  ungroup() %>% 
  group_by(hour) %>% 
  summarise(wape=100*sum(abs(obs.pct.h-sim.pct.h))/sum(obs.pct.h))

error_wape %>% 
  ggplot(aes(x=hour, y=wape)) +
  geom_line() +
  theme(
    legend.position = "bottom",
    legend.title = element_blank(),
    text = element_text(color = "#22211d"),
    plot.background = element_rect(fill = "#f5f5f2", color = NA), 
    panel.background = element_rect(fill = "#f5f5f2", color = NA), 
    legend.background = element_rect(fill = "#f5f5f2", color = NA)
  ) +
  ylab("Weighted Absolute Error (%)") +
  xlab("Hour of the day")

ggsave("car_wape.png", width= 5, height = 3)

carDataPct.h %>% 
  group_by(hour) %>% 
  summarise(obs.pct.h.avg=mean(obs.pct.h),
            sim.pct.h.avg=mean(sim.pct.h)) %>% 
  ggplot(aes(x=hour))+
  geom_line(aes(y=obs.pct.h.avg, color="observation")) + 
  geom_line(aes(y=sim.pct.h.avg, color="simulation"))

ggsave(paste0(outputDir,"carpctavg.png"), width= 5, height = 5)
